Separate messages for missing PAT and insufficent permissions
