+++ tests/Plugins/GitHubModule/StandardQueryRequirements/License_UnitTest.py
@@ -12,18 +12,6 @@ from RepoAuditor.Plugins.GitHub.StandardRequirements.License import License
 from RepoAuditor.Requirement import EvaluateResult
 
 
-@pytest.fixture(name="query_data")
-def query_data_fixture(session):
-    return {
-        "session": session,
-        "standard": {
-            "license": {
-                "name": "MIT License",
-            },
-        },
-    }
-
-
 @pytest.fixture(name="requirement")
 def requirement_fixture():
     return License()
