+++ tests/Plugins/GitHubModule/StandardQueryRequirements/MergeCommitMessage_UnitTest.py
@@ -12,17 +12,6 @@ from RepoAuditor.Plugins.GitHub.StandardRequirements.MergeCommitMessage import M
 from RepoAuditor.Requirement import EvaluateResult
 
 
-@pytest.fixture(name="query_data")
-def query_data_fixture(session):
-    return {
-        "session": session,
-        "standard": {
-            "allow_merge_commit": True,
-            "merge_commit_message": "BLANK",
-        },
-    }
-
-
 @pytest.fixture(name="requirement")
 def requirement_fixture():
     return MergeCommitMessage()
