+++ tests/Plugins/GitHubModule/StandardQueryRequirements/SecretScanning_UnitTest.py
@@ -12,20 +12,6 @@ from RepoAuditor.Plugins.GitHub.StandardRequirements.SecretScanning import Secre
 from RepoAuditor.Requirement import EvaluateResult
 
 
-@pytest.fixture(name="query_data")
-def query_data_fixture(session):
-    return {
-        "session": session,
-        "standard": {
-            "security_and_analysis": {
-                "secret_scanning": {
-                    "status": True,
-                },
-            },
-        },
-    }
-
-
 @pytest.fixture(name="requirement")
 def requirement_fixture():
     return SecretScanning()
